Refactored article table code
